libcontainer: Check if pid path is valid before unlink
If the pid path is NULL, don't call unlink on it.  This removes an extra
error log generated when unlink returns -EINVAL.

BUG=none
TEST=strace with invalid config, check no EINVAL from unlink

Reviewed-on: https://chromium-review.googlesource.com/347510

+++ libcontainer/libcontainer.c
@@ -690,7 +690,8 @@ int container_start(struct container *c, const struct container_config *config)
 error_rmdir:
 	umount(c->runfsroot);
 	rmdir(c->runfsroot);
-	unlink(c->pid_file_path);
+	if (c->pid_file_path)
+		unlink(c->pid_file_path);
 	free(c->pid_file_path);
 	rmdir(c->runfs);
 	free(c->runfsroot);
@@ -717,7 +718,7 @@ static int container_teardown(struct container *c)
 		ret = -errno;
 	if (rmdir(c->runfsroot))
 		ret = -errno;
-	if (unlink(c->pid_file_path))
+	if (c->pid_file_path && unlink(c->pid_file_path))
 		ret = -errno;
 	if (rmdir(c->runfs))
 		ret = -errno;
